feat: US-05 - SIMD horizontal sum deduplication
Create simd_native/reduction.rs with canonical hsum_avx256 and
hsum_avx512 helpers. Replace 10+ duplicate horizontal reduction
sequences across x86_avx2/dot.rs, x86_avx2/l2.rs,
x86_avx2_similarity.rs, and adc.rs with calls to the shared helper.
